requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
db-migrator==1.1.0
discord==1.0.1
discord.py==1.4.1
Flask==1.1.2
html5lib==1.1
idna==2.10
idna-ssl==1.1.0
influxdb==5.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mechanize==0.4.5
msgpack==0.6.1
multidict==4.7.6
praw==7.1.0
prawcore==1.5.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.25.10
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
